sdm845-common: configs: Add Adaptive Battery CPU hints for crosshatch
Add hints allowing Adaptive Battery to limit the frequencies of the
big and little cores to 50%, 60%, 70%, 80%, or 90%. This is to make
them consistent with the hints on redbull devices (see ag/12036507).

In particular:
 - Add hints for the big and little cores.
 - Make the 50% and 60% hints override the min frequency of the
   INTERACTION hint.
 - Make the AUDIO_LAUNCH also set a maximum cpu frequency to prevent
   situations where the max frequency is smaller than the min
   frequency.

See go/powerbrain-cpu-crosshatch

Bug: 151080800
Bug: 166599374
Test: Make and flash, use a dev version of Turbo to set a hint

Change-Id: I7ba89d9af8836b6997dd5edfb571a18853489043
(cherry picked from commit 2d470c895bfa4229621ce53e8f41edc5892e8834)
[@jjpprrrr: Since we tweaked INTERACTION min frequency before, adapt
accordingly to make 50% and 60% hint override the min frequency. The
default indexes are also changed to 4 because we have an extra value
in the list.]
Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Signed-off-by: DarkAngelGR <asavvo01@gmail.com>